feat: fix comments

+++ crates/walrus-service/src/mapping.rs
@@ -91,8 +91,11 @@ fn is_rotation(pairs: &[SliverPair]) -> bool {
 /// * `total_shards` - The total number of shards in the system.
 /// * `blob_id` - The Blob ID that produced the sliver. It is interpreted as a big-endian unsigned
 /// integer, and then used to compute the offset for the sliver pair index.
-pub fn shard_index_for_pair(pair_idx: usize, total_shards: usize, blob_id: &BlobId) -> ShardIndex {
-    let index = (bytes_mod(blob_id.as_ref(), total_shards) + pair_idx) % total_shards;
+///
+/// # Panics
+/// Panic if the total number of shards is greater than `u16::MAX`.
+pub fn shard_index_for_pair(pair_idx: u16, total_shards: usize, blob_id: &BlobId) -> ShardIndex {
+    let index = (bytes_mod(blob_id.as_ref(), total_shards) + pair_idx as usize) % total_shards;
     ShardIndex(index as u16)
 }
 
@@ -221,7 +224,7 @@ mod tests {
     fn test_shard_index_for_pair(
         total_shards: usize,
         blob_id_value: u64,
-        pair_idx: usize,
+        pair_idx: u16,
         shard_idx: ShardIndex,
     ) {
         let blob_id = test_utils::blob_id_from_u64(blob_id_value);
